+++ src/UI/SignTransactionsModals/SignWithDeviceModal/components/SignStepBody.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Address } from '@multiversx/sdk-core/out';
+import { Address } from '@multiversx/sdk-core';
 import classNames from 'classnames';
 
 import { withStyles, WithStylesImportType } from 'hocs/withStyles';