+++ src/hooks/login/useIframeLogin.ts
@@ -9,7 +9,7 @@ import { setAccountProvider } from 'providers/accountProvider';
 import { loginAction } from 'reduxStore/commonActions';
 import { useDispatch, useSelector } from 'reduxStore/DappProviderContext';
 import { networkSelector } from 'reduxStore/selectors/networkConfigSelectors';
-import { setAccount } from 'reduxStore/slices';
+import { emptyAccount, setAccount, setAddress } from 'reduxStore/slices';
 import {
   LoginHookGenericStateType,
   LoginMethodsEnum,
@@ -52,6 +52,9 @@ export const useIframeLogin = ({
     clearInitiatedLogins({
       skipLoginMethod: LoginMethodsEnum.iframe
     });
+
+    dispatch(setAddress(emptyAccount.address));
+    dispatch(setAccount(emptyAccount));
 
     setIsLoading(true);
     const provider = IframeProvider.getInstance();