+++ src/hooks/login/useLedgerLogin.ts
@@ -7,6 +7,9 @@ import { loginAction } from 'reduxStore/commonActions';
 import { useDispatch, useSelector } from 'reduxStore/DappProviderContext';
 import { ledgerAccountSelector } from 'reduxStore/selectors';
 import {
+  emptyAccount,
+  setAccount,
+  setAddress,
   setLedgerAccount,
   setLedgerLogin,
   updateLedgerAccount
@@ -281,6 +284,8 @@ export const useLedgerLogin = ({
     }
 
     clearInitiatedLogins();
+    dispatch(setAddress(emptyAccount.address));
+    dispatch(setAccount(emptyAccount));
     setError('');
 
     try {