+++ src/hooks/login/useMetamaskLogin.ts
@@ -5,6 +5,7 @@ import { SECOND_LOGIN_ATTEMPT_ERROR } from 'constants/errorsMessages';
 import { setAccountProvider } from 'providers/accountProvider';
 import { loginAction } from 'reduxStore/commonActions';
 import { useDispatch } from 'reduxStore/DappProviderContext';
+import { emptyAccount, setAccount, setAddress } from 'reduxStore/slices';
 import {
   InitiateLoginFunctionType,
   LoginHookGenericStateType,
@@ -44,6 +45,8 @@ export const useMetamaskLogin = ({
     }
 
     clearInitiatedLogins();
+    dispatch(setAddress(emptyAccount.address));
+    dispatch(setAccount(emptyAccount));
 
     setIsLoading(true);
     const provider: MetamaskProvider = MetamaskProvider.getInstance();