+++ src/hooks/login/useWebWalletLogin.ts
@@ -2,7 +2,13 @@ import { useState } from 'react';
 import { SECOND_LOGIN_ATTEMPT_ERROR } from 'constants/errorsMessages';
 import { useDispatch, useSelector } from 'reduxStore/DappProviderContext';
 import { networkSelector } from 'reduxStore/selectors';
-import { setWalletLogin, setCustomWalletAddress } from 'reduxStore/slices';
+import {
+  setWalletLogin,
+  setCustomWalletAddress,
+  setAccount,
+  emptyAccount,
+  setAddress
+} from 'reduxStore/slices';
 import { newWalletProvider } from 'utils';
 import { getIsLoggedIn } from 'utils/getIsLoggedIn';
 import { getWindowLocation } from 'utils/window/getWindowLocation';
@@ -59,6 +65,8 @@ export const useWebWalletLogin = ({
       throw new Error(SECOND_LOGIN_ATTEMPT_ERROR);
     }
 
+    dispatch(setAddress(emptyAccount.address));
+    dispatch(setAccount(emptyAccount));
     dispatch(setCustomWalletAddress(customWalletAddress));
 
     try {