+++ src/services/transactions/isCrossShardTransaction.ts
@@ -1,4 +1,4 @@
-import { Address } from '@multiversx/sdk-core/out';
+import { Address } from '@multiversx/sdk-core';
 
 import { getShardOfAddress } from 'utils/account/getShardOfAddress';